+++ composeApp/src/commonMain/kotlin/tech/kotlinlang/permission/ui/component/PdfViewerScreen.kt
@@ -1,0 +1,6 @@
+package tech.kotlinlang.permission.ui.component
+
+import androidx.compose.runtime.Composable
+
+@Composable
+expect fun PdfViewerScreen()