+++ composeApp/src/commonMain/kotlin/tech/kotlinlang/permission/ui/permission/PermissionsScreen.kt
@@ -1,0 +1,8 @@
+package tech.kotlinlang.permission.ui.permission
+
+import androidx.compose.runtime.Composable
+
+@Composable
+expect fun PermissionsScreen(
+    onCameraClick: () -> Unit,
+)