+++ composeApp/src/wasmJsMain/kotlin/tech/kotlinlang/permission/ui/camera/CameraScreen.wasmJs.kt
@@ -1,0 +1,25 @@
+package tech.kotlinlang.permission.ui.camera
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.LocalTextStyle
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.style.TextAlign
+
+@Composable
+actual fun CameraScreen() {
+    Box(
+        modifier = Modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        Text(
+            text = "Camera Screen\nComing Soon!",
+            style = LocalTextStyle.current.copy(
+                textAlign = TextAlign.Center,
+            )
+        )
+    }
+}